Add : mypage 내부 컴포넌트 생성
