+++ src/components/MyPage/Item/UserPlanListItem.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>여행 계획</div>
+</template>
+
+<script setup>
+
+</script>
+
+<style scoped>
+
+</style>
