+++ src/components/MyPage/UserPlanList.vue
@@ -1,0 +1,12 @@
+<script setup>
+import UserPlanListItem from "./Item/UserPlanListItem.vue";
+
+</script>
+
+<template>
+    <user-plan-list-item></user-plan-list-item>
+</template>
+
+<style scoped>
+
+</style>
