+++ src/components/MyPage/UserProfile.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>프로필</div>
+</template>
+
+<script setup>
+
+</script>
+
+<style scoped>
+
+</style>
