+++ src/components/MyPage/UserReview.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>여행 후기</div>
+</template>
+
+<script setup>
+
+</script>
+
+<style scoped>
+
+</style>
